@@ -1,4 +1,5 @@
 <?php
+
 namespace SimpleSoftwareIO\SMS\Drivers;
 
 use SimpleSoftwareIO\SMS\IncomingMessage;
@@ -6,11 +7,15 @@ use SimpleSoftwareIO\SMS\SMSNotSentException;
 
 abstract class AbstractSMS
 {
+
+    protected $debug;
+
     /**
-     * Throw a not sent exception
+     * Throw a not sent exception.
      *
-     * @param string $message
-     * @param null|integer $code
+     * @param string   $message
+     * @param null|int $code
+     *
      * @throws SMSNotSentException
      */
     protected function throwNotSentException($message, $code = null)
@@ -65,4 +70,16 @@ abstract class AbstractSMS
      * @return mixed
      */
     abstract protected function processReceive($rawMessage);
+
+    /**
+     * Defines if debug is enabled or disabled (SMS77)
+     *
+     * @param $debug
+     */
+    public function setDebug($debug)
+    {
+        $this->debug = $debug;
+    }
+
+
 }
