@@ -1,9 +1,10 @@
 <?php
+
 namespace SimpleSoftwareIO\SMS\Drivers;
 
 use GuzzleHttp\Client;
-use SimpleSoftwareIO\SMS\MakesRequests;
 use SimpleSoftwareIO\SMS\DoesNotReceive;
+use SimpleSoftwareIO\SMS\MakesRequests;
 use SimpleSoftwareIO\SMS\OutgoingMessage;
 
 class LabsMobileSMS extends AbstractSMS implements DriverInterface
@@ -45,7 +46,7 @@ class LabsMobileSMS extends AbstractSMS implements DriverInterface
 
         foreach ($message->getTo() as $to) {
             $data = [
-                'msisdn' => $to,
+                'msisdn'  => $to,
                 'message' => $composeMessage,
             ];
 
