@@ -1,4 +1,5 @@
 <?php
+
 namespace SimpleSoftwareIO\SMS\Drivers;
 
 use Plivo\RestAPI as Plivo;
@@ -14,7 +15,7 @@ class PlivoSMS extends AbstractSMS implements DriverInterface
     protected $plivo;
 
     /**
-     * The authentication token
+     * The authentication token.
      */
     protected $authToken;
 
@@ -26,7 +27,7 @@ class PlivoSMS extends AbstractSMS implements DriverInterface
      */
     public function __construct($authId, $authToken)
     {
-        $this->plivo = new Plivo( $authId, $authToken );
+        $this->plivo = new Plivo($authId, $authToken);
         $this->authToken = $authToken;
     }
 
@@ -42,8 +43,8 @@ class PlivoSMS extends AbstractSMS implements DriverInterface
 
         foreach ($message->getTo() as $to) {
             $response = $this->plivo->send_message([
-                'dst' => $to,
-                'src' => $from,
+                'dst'  => $to,
+                'src'  => $from,
                 'text' => $composeMessage,
             ]);
 
@@ -82,7 +83,7 @@ class PlivoSMS extends AbstractSMS implements DriverInterface
 
         $rawMessages = $this->plivo->get_messages([
             'offset' => $start,
-            'limit' => $end,
+            'limit'  => $end,
         ]);
 
         $incomingMessages = [];
@@ -147,8 +148,7 @@ class PlivoSMS extends AbstractSMS implements DriverInterface
         $signature = $_SERVER['X-Plivo-Signature'];
         $authToken = $this->authToken;
 
-        if ( ! $this->plivo->validate_signature( $url, $data, $signature, $authToken )  )
-        {
+        if (!$this->plivo->validate_signature($url, $data, $signature, $authToken)) {
             throw new \InvalidArgumentException('This request was not able to verify it came from Plivo.');
         }
 
