@@ -1,4 +1,7 @@
 <?php
+
 namespace SimpleSoftwareIO\SMS;
 
-class SMSNotSentException extends \Exception {}
+class SMSNotSentException extends \Exception
+{
+}
